Restructure the BF16 kernel curriculum
