+++ book/src/SUMMARY.md
@@ -19,12 +19,13 @@
     - [🚧 Quantized Matvec](./week2-03-quantized-matvec.md)
     - [🚧 Decode Attention](./week2-04-decode-attention.md)
     - [🚧 Fast Kernels](./week2-05-fast-kernels.md)
+    - [🚧 SIMD-Matrix Prefill](./week2-06-simd-matrix-prefill.md)
 - [🚧 Week 3: Build a Mini vLLM](./week3-overview.md)
     - [🚧 Continuous Batching](./week3-01-continuous-batching.md)
-    - [🚧 FlashAttention for Prefill](./week3-02-flash-attention.md)
-    - [🚧 Chunked Prefill](./week3-03-chunked-prefill.md)
-    - [🚧 Paged Attention, Part 1](./week3-04-paged-attention-part1.md)
-    - [🚧 Paged Attention, Part 2](./week3-05-paged-attention-part2.md)
+    - [🚧 Chunked Prefill](./week3-02-chunked-prefill.md)
+    - [🚧 Paged Attention, Part 1](./week3-03-paged-attention-part1.md)
+    - [🚧 Paged Attention, Part 2](./week3-04-paged-attention-part2.md)
+    - [🚧 Paged FlashAttention](./week3-05-flash-attention.md)
     - [Optional: Mixture of Experts](./week3-optional-moe.md)
     - [🚧 Optional: Performance Lab](./week3-performance-lab.md)
     - [🚧 Optional: Speculative Decoding](./week3-optional-speculative-decoding.md)
