+++ main.py
@@ -21,15 +21,6 @@ parser.add_argument("--sampler-temp", type=float, default=0)
 parser.add_argument("--sampler-top-p", type=float, default=None)
 parser.add_argument("--sampler-top-k", type=int, default=None)
 parser.add_argument("--enable-thinking", action="store_true")
-flash_group = parser.add_mutually_exclusive_group()
-flash_group.add_argument(
-    "--enable-flash-attn", dest="enable_flash_attn", action="store_true"
-)
-flash_group.add_argument(
-    "--disable-flash-attn", dest="enable_flash_attn", action="store_false"
-)
-parser.set_defaults(enable_flash_attn=None)
-parser.add_argument("--enable-performance-lab", action="store_true")
 parser.add_argument(
     "--disable-paged-attention",
     action="store_true",
@@ -44,23 +35,29 @@ parser.add_argument(
         "rmsnorm",
         "rope",
         "swiglu",
+        "simd-matmul",
     ),
     help="run one cumulative Week 2 model checkpoint",
 )
 
 args = parser.parse_args()
 
-if (
-    args.solution != "mlx"
-    and args.loader == "week3"
-    and args.enable_flash_attn is not False
-    and args.device != "gpu"
-):
-    parser.error("Week 3 FlashAttention requires --device gpu")
 if args.week2_checkpoint is not None and args.loader != "week2":
     parser.error("--week2-checkpoint requires --loader week2")
 if args.week2_checkpoint is not None and args.solution == "mlx":
     parser.error("--week2-checkpoint is not supported with --solution mlx")
+if (
+    args.solution != "mlx"
+    and args.device != "gpu"
+    and (
+        args.loader == "week3"
+        or (args.loader == "week2" and args.week2_checkpoint != "kv-cache")
+    )
+):
+    parser.error(
+        "The completed Week 2 and Week 3 custom-kernel models are GPU-only; "
+        "use the Week 2 kv-cache checkpoint for the readable pre-kernel path"
+    )
 if args.disable_paged_attention and args.loader != "week3":
     parser.error("--disable-paged-attention requires --loader week3")
 if args.disable_paged_attention and args.solution == "mlx":
@@ -112,20 +109,12 @@ with mx.stream(mx.gpu if args.device == "gpu" else mx.cpu):
         tiny_llm_model = mlx_model
     else:
         if args.loader == "week1":
-            if args.enable_flash_attn:
-                print("--enable-flash-attn belongs to Week 3; ignoring it")
-            if args.enable_performance_lab:
-                print("--enable-performance-lab belongs to Week 3; ignoring it")
             print(f"Using week1 loader for {args.model}")
             tiny_llm_model = models.dispatch_model(args.model, mlx_model, week=1)
         elif args.loader == "week2":
             print(
                 f"Using week2 loader with thinking={args.enable_thinking} for {args.model}"
             )
-            if args.enable_flash_attn:
-                print("--enable-flash-attn belongs to Week 3; ignoring it")
-            if args.enable_performance_lab:
-                print("--enable-performance-lab belongs to Week 3; ignoring it")
             dispatch_kwargs = {}
             if args.week2_checkpoint is not None:
                 dispatch_kwargs["checkpoint"] = args.week2_checkpoint
@@ -144,16 +133,13 @@ with mx.stream(mx.gpu if args.device == "gpu" else mx.cpu):
                 draft_tiny_llm_model = None
         elif args.loader == "week3":
             print(
-                f"Using week3 loader with flash_attn={args.enable_flash_attn} "
-                f"paged_attention={not args.disable_paged_attention} "
+                f"Using week3 loader with paged_attention={not args.disable_paged_attention} "
                 f"thinking={args.enable_thinking} for {args.model}"
             )
             tiny_llm_model = models.dispatch_model(
                 args.model,
                 mlx_model,
                 week=3,
-                enable_flash_attn=args.enable_flash_attn,
-                enable_performance_lab=args.enable_performance_lab,
                 enable_paged_attention=not args.disable_paged_attention,
             )
             if draft_mlx_model is not None:
@@ -162,8 +148,6 @@ with mx.stream(mx.gpu if args.device == "gpu" else mx.cpu):
                     args.draft_model,
                     draft_mlx_model,
                     week=3,
-                    enable_flash_attn=args.enable_flash_attn,
-                    enable_performance_lab=args.enable_performance_lab,
                     enable_paged_attention=not args.disable_paged_attention,
                 )
             else:
