+++ src/extensions_ref/src/tiny_llm_ext.h
@@ -131,37 +131,6 @@ private:
     int num_kv_heads_;
 };
 
-mx::array flash_attention(const mx::array &q, const mx::array &k, const mx::array &v, const mx::array &mask,
-                          const float scale, const int mask_mode, const int num_kv_heads, const int num_heads,
-                          mx::StreamOrDevice s = {});
-
-class FlashAttention : public mx::Primitive {
-public:
-    explicit FlashAttention(mx::Stream stream, const float scale, const int mask_mode, const int num_kv_heads,
-                            const int num_heads)
-        : mx::Primitive(stream),
-          scale_(scale),
-          mask_mode_(mask_mode),
-          num_kv_heads_(num_kv_heads),
-          num_heads_(num_heads) {};
-
-    void eval_cpu(const std::vector<mx::array> &inputs, std::vector<mx::array> &outputs) override;
-    void eval_gpu(const std::vector<mx::array> &inputs, std::vector<mx::array> &outputs) override;
-
-    std::pair<std::vector<mx::array>, std::vector<int>> vmap(const std::vector<mx::array> &inputs,
-                                                             const std::vector<int> &axes) override {
-        throw std::runtime_error("FlashAttention has no vmap implementation.");
-    }
-
-    const char *name() const override { return "FlashAttention"; }
-
-private:
-    float scale_;
-    int mask_mode_;
-    int num_kv_heads_;
-    int num_heads_;
-};
-
 mx::array paged_attention(const mx::array &q, const mx::array &key_pages, const mx::array &value_pages,
                           const mx::array &block_table, const mx::array &context_lens, const float scale,
                           const bool is_causal, const int num_kv_heads, const int num_heads, mx::StreamOrDevice s = {});
