+++ src/tiny_llm/qwen3_week3.py
@@ -21,7 +21,6 @@ class Qwen3MultiHeadAttention:
         max_seq_len: int = 32768,
         theta: int = 1000000,
         rms_norm_eps: float = 1e-5,
-        use_flash_attention: bool = False,
         use_paged_attention: bool = True,
     ):
         pass
@@ -70,7 +69,6 @@ class Qwen3TransformerBlock:
         mlp: Qwen3MLP | Moe,
         max_seq_len: int = 32768,
         theta: int = 1000000,
-        use_flash_attention: bool = False,
         use_paged_attention: bool = True,
     ):
         pass
@@ -94,8 +92,6 @@ class Qwen3ModelWeek3:
         self,
         mlx_model: Any,
         page_size: int = 128,
-        enable_flash_attn: bool | None = None,
-        enable_performance_lab: bool = False,
         enable_paged_attention: bool = True,
     ):
         self.num_hidden_layers = mlx_model.args.num_hidden_layers
