+++ src/tiny_llm_ref/quantize.py
@@ -30,9 +30,10 @@ class QuantizedWeights:
         use_simdgroup_matmul: bool = False,
         use_simdgroup_matvec: bool = True,
     ) -> "QuantizedWeights":
+        biases = mlx_layer.biases
         return QuantizedWeights(
-            scales=mlx_layer.scales,
-            biases=mlx_layer.biases,
+            scales=mlx_layer.scales.astype(mx.bfloat16),
+            biases=None if biases is None else biases.astype(mx.bfloat16),
             group_size=mlx_layer.group_size,
             bits=mlx_layer.bits,
             weight=mlx_layer.weight,
@@ -91,7 +92,7 @@ def dequantize_linear(mx_layer: Any) -> mx.array:
         mx_layer.biases,
         mx_layer.group_size,
         mx_layer.bits,
-    )
+    ).astype(mx.bfloat16)
 
 
 def dequantize_weights(
