+++ tests_refsol/test_week_2_day_4.py
@@ -25,17 +25,18 @@ def test_model_integrates_decode_attention_before_fast_kernels():
 
 
 def test_fast_attention_matches_grouped_attention():
-    query = mx.random.normal((2, 4, 3, 16)).astype(mx.float32)
-    key = mx.random.normal((2, 2, 5, 16)).astype(mx.float32)
-    value = mx.random.normal((2, 2, 5, 16)).astype(mx.float32)
+    query = mx.random.normal((2, 4, 3, 16)).astype(mx.bfloat16)
+    key = mx.random.normal((2, 2, 5, 16)).astype(mx.bfloat16)
+    value = mx.random.normal((2, 2, 5, 16)).astype(mx.bfloat16)
     mask = mx.broadcast_to(
-        mx.array([0, 0, 0, 0, -mx.inf], dtype=mx.float32), (2, 1, 3, 5)
+        mx.array([0, 0, 0, 0, -mx.inf], dtype=mx.bfloat16), (2, 1, 3, 5)
     )
     scale = 16**-0.5
     result = scaled_dot_product_attention(query, key, value, scale, mask)
     expected = scaled_dot_product_attention_grouped(query, key, value, scale, mask)
     assert result.shape == query.shape
-    assert_allclose(result, expected, mx.float32, atol=1e-5, rtol=1e-5)
+    assert result.dtype == mx.bfloat16
+    assert_allclose(result, expected, mx.bfloat16, atol=2e-2, rtol=2e-2)
 
 
 def test_custom_metal_attention_matches_grouped_attention():
