+++ tests_refsol/test_week_2_day_6.py
@@ -1,7 +1,5 @@
-"""Week 3 performance-lab tests."""
+"""Week 2 Day 6 SIMD-matrix prefill tests."""
 
-import importlib
-import inspect
 import mlx.core as mx
 import pytest
 
@@ -9,7 +7,6 @@ from .tiny_llm_base import (
     QuantizedEmbedding,
     QuantizedWeights,
     Qwen3ModelWeek2,
-    Qwen3ModelWeek3,
     quantized_matmul,
     quantized_matmul_vanilla,
 )
@@ -18,16 +15,21 @@ from .utils import (
     qwen3_0_6b_model_exists,
     qwen3_1_7b_model_exists,
     qwen3_4b_model_exists,
+    tiny_qwen3_mlx_model,
 )
 from mlx_lm import load
 
 
-def test_week2_model_does_not_depend_on_week3_flash_attention():
-    module = importlib.import_module(Qwen3ModelWeek2.__module__)
-    assert "flash_attention" not in inspect.getsource(module)
+def test_simd_matmul_checkpoint_is_completed_week2_model():
+    model = Qwen3ModelWeek2(tiny_qwen3_mlx_model(), checkpoint="simd-matmul")
+    layer = model.layers_inner[0]
+
+    assert model.embedding.use_custom_kernel
+    assert model.embedding.weight.use_simdgroup_matmul
+    assert layer.self_attn.wq.use_simdgroup_matmul
 
 
-def test_performance_lab_simdgroup_matmul_matches_vanilla_gpu():
+def test_task_2_simdgroup_matmul_matches_vanilla_gpu():
     with mx.stream(mx.gpu):
         inputs = mx.random.normal((128, 256)).astype(mx.bfloat16)
         weight = mx.random.normal((96, 256)).astype(mx.bfloat16)
@@ -48,7 +50,7 @@ def test_performance_lab_simdgroup_matmul_matches_vanilla_gpu():
         assert_allclose(tiled, vanilla, mx.bfloat16, atol=1.0, rtol=2e-2)
 
 
-def test_performance_lab_simdgroup_matmul_uses_accurate_partial_tiles_gpu():
+def test_task_2_simdgroup_matmul_uses_accurate_partial_tiles_gpu():
     """A non-multiple-of-eight prefill must not accumulate in bfloat16."""
     with mx.stream(mx.gpu):
         inputs = mx.random.normal((10, 256)).astype(mx.bfloat16)
@@ -70,7 +72,7 @@ def test_performance_lab_simdgroup_matmul_uses_accurate_partial_tiles_gpu():
         assert_allclose(tiled, vanilla, mx.bfloat16, atol=0.25, rtol=1e-2)
 
 
-def test_performance_lab_custom_embedding_matches_readable_path():
+def test_task_4_custom_embedding_matches_readable_path():
     weight = mx.random.normal((17, 256)).astype(mx.bfloat16)
     packed, scales, biases = mx.quantize(weight, group_size=128, bits=4)
     quantized = QuantizedWeights(scales, biases, 128, 4, packed)
@@ -84,9 +86,6 @@ def test_performance_lab_custom_embedding_matches_readable_path():
         atol=2e-2,
         rtol=2e-2,
     )
-
-
-# TODO: task 1 tests
 
 
 @pytest.mark.skipif(
@@ -108,9 +107,9 @@ def test_utils_qwen3_1_7b():
     pass
 
 
-def helper_test_task_3(model_name: str, iters: int = 10):
+def helper_test_task_6(model_name: str, iters: int = 10):
     mlx_model, tokenizer = load(model_name)
-    model = Qwen3ModelWeek3(mlx_model, enable_performance_lab=True)
+    model = Qwen3ModelWeek2(mlx_model, checkpoint="simd-matmul")
     assert model.embedding.use_custom_kernel
     assert model.embedding.weight.use_simdgroup_matmul
     assert all(layer.self_attn.wq.use_simdgroup_matmul for layer in model.layers_inner)
@@ -131,29 +130,29 @@ def helper_test_task_3(model_name: str, iters: int = 10):
 @pytest.mark.skipif(
     not qwen3_0_6b_model_exists(), reason="Qwen3-0.6B-4bit model not found"
 )
-def test_task_3_qwen3_0_6b():
-    helper_test_task_3("Qwen/Qwen3-0.6B-MLX-4bit", 5)
+def test_task_6_qwen3_0_6b():
+    helper_test_task_6("Qwen/Qwen3-0.6B-MLX-4bit", 5)
 
 
 @pytest.mark.skipif(not qwen3_4b_model_exists(), reason="Qwen3-4B-4bit model not found")
-def test_task_3_qwen3_4b():
-    helper_test_task_3("Qwen/Qwen3-4B-MLX-4bit", 1)
+def test_task_6_qwen3_4b():
+    helper_test_task_6("Qwen/Qwen3-4B-MLX-4bit", 1)
 
 
 @pytest.mark.skipif(
     not qwen3_1_7b_model_exists(), reason="Qwen3-1.7B-4bit model not found"
 )
-def test_task_3_qwen3_1_7b():
-    helper_test_task_3("Qwen/Qwen3-1.7B-MLX-4bit", 3)
+def test_task_6_qwen3_1_7b():
+    helper_test_task_6("Qwen/Qwen3-1.7B-MLX-4bit", 3)
 
 
-def helper_test_task_4(
+def helper_test_task_6_incremental(
     model_name: str,
     seq_len: int,
     iters: int = 1,
 ):
     mlx_model, tokenizer = load(model_name)
-    model = Qwen3ModelWeek3(mlx_model, enable_performance_lab=True)
+    model = Qwen3ModelWeek2(mlx_model, checkpoint="simd-matmul")
     for _ in range(iters):
         inputs = mx.random.randint(0, tokenizer.vocab_size, (1, seq_len))
         ref_outputs = mlx_model(inputs)
@@ -175,13 +174,13 @@ def helper_test_task_4(
 @pytest.mark.skipif(
     not qwen3_0_6b_model_exists(), reason="Qwen3-0.6B-4bit model not found"
 )
-def test_task_4_qwen3_0_6b():
-    helper_test_task_4("Qwen/Qwen3-0.6B-MLX-4bit", seq_len=3)
+def test_task_6_incremental_qwen3_0_6b():
+    helper_test_task_6_incremental("Qwen/Qwen3-0.6B-MLX-4bit", seq_len=3)
 
 
 @pytest.mark.skipif(not qwen3_4b_model_exists(), reason="Qwen3-4B-4bit model not found")
-def test_task_4_qwen3_4b():
-    helper_test_task_4(
+def test_task_6_incremental_qwen3_4b():
+    helper_test_task_6_incremental(
         "Qwen/Qwen3-4B-MLX-4bit",
         seq_len=3,
     )
@@ -190,24 +189,21 @@ def test_task_4_qwen3_4b():
 @pytest.mark.skipif(
     not qwen3_1_7b_model_exists(), reason="Qwen3-1.7B-4bit model not found"
 )
-def test_task_4_qwen3_1_7b():
-    helper_test_task_4("Qwen/Qwen3-1.7B-MLX-4bit", seq_len=3)
+def test_task_6_incremental_qwen3_1_7b():
+    helper_test_task_6_incremental("Qwen/Qwen3-1.7B-MLX-4bit", seq_len=3)
 
 
 class FakeEmbedding:
     def __call__(self, inputs):
-        return mx.stack([inputs, inputs + 1], axis=-1).astype(mx.float32)
+        return mx.stack([inputs, inputs + 1], axis=-1).astype(mx.bfloat16)
 
     def as_linear(self, hidden):
         return hidden
 
 
-@pytest.mark.parametrize("model_class", [Qwen3ModelWeek2, Qwen3ModelWeek3])
 @pytest.mark.parametrize("logits_to_keep,expected_length", [(1, 1), (None, 4)])
-def test_logits_to_keep_controls_output_length(
-    model_class, logits_to_keep, expected_length
-):
-    model = model_class.__new__(model_class)
+def test_task_5_logits_to_keep_controls_output_length(logits_to_keep, expected_length):
+    model = Qwen3ModelWeek2.__new__(Qwen3ModelWeek2)
     model.num_hidden_layers = 0
     model.embedding = FakeEmbedding()
     model.layers_inner = []
