+++ tests_refsol/test_week_3_day_2.py
@@ -1,206 +1,68 @@
-"""Week 3 Day 2 FlashAttention tests."""
+"""Week 3 Day 2 chunked-prefill tests."""
 
-import pytest
-import time
 import mlx.core as mx
-from .tiny_llm_base import *
-from .utils import *
+import pytest
+
+from .tiny_llm_base import Request, TinyKvFullCache
 
 
-def attention_helper(
-    stream: mx.Stream,
-    H_q,
-    H,
-    L,
-    E,
-    S,
-    BATCH,
-    mask_mode: str,
-    precision=mx.float32,
-):
-    with mx.stream(stream):
-        q_shape = (BATCH, H_q, L, E)
-        kv_shape = (BATCH, H, S, E)
-        mask_shape = (BATCH, H_q, L, S)
-        scale = 0.9
-        for _ in range(100):
-            query = mx.random.uniform(shape=q_shape, dtype=precision)
-            key = mx.random.uniform(shape=kv_shape, dtype=precision)
-            value = mx.random.uniform(shape=kv_shape, dtype=precision)
-            if mask_mode == "no_mask":
-                mask = None
-            elif mask_mode == "mask":
-                mask = mx.random.uniform(shape=mask_shape, dtype=precision)
-            elif mask_mode == "causal":
-                mask = "causal"
-            else:
-                raise ValueError(f"Unknown mask_mode: {mask_mode}")
+class FakeDetokenizer:
+    def __init__(self, _):
+        self.text = ""
 
-            reference_output = mx.fast.scaled_dot_product_attention(
-                q=query,
-                k=key,
-                v=value,
-                scale=scale,
-                mask=mask,
-            )
-            user_output = flash_attention(
-                query,
-                key,
-                value,
-                scale=scale,
-                mask=mask,
-            )
-            mx.eval(user_output)  # so that any error will be caught here
-            assert user_output.dtype == precision
-            comparison_precision = (
-                mx.bfloat16 if precision == mx.bfloat16 else mx.float16
-            )
-            assert_allclose(
-                user_output, reference_output, precision=comparison_precision
-            )
+    def add_token(self, token):
+        self.text += str(token)
 
 
-def time_flash_attention(
-    stream: mx.Stream,
-    query: mx.array,
-    key: mx.array,
-    value: mx.array,
-    scale: float,
-    mask: mx.array | str,
-    num_iters: int = 4,
-) -> float:
-    with mx.stream(stream):
-        start = time.perf_counter()
-        for _ in range(num_iters):
-            output = flash_attention(query, key, value, scale=scale, mask=mask)
-            mx.eval(output)
-    return (time.perf_counter() - start) / num_iters
+class FakeTokenizer:
+    eos_token_id = 99
+    _tokenizer = object()
+    detokenizer = FakeDetokenizer(_tokenizer)
+
+    def encode(self, prompt, add_special_tokens=False):
+        assert not add_special_tokens
+        return list(range(1, len(prompt) + 1))
 
 
-def median(values: list[float]) -> float:
-    values = sorted(values)
-    return values[len(values) // 2]
+class FakeModel:
+    num_hidden_layers = 1
+
+    def __init__(self):
+        self.calls = []
+
+    def create_kv_cache(self):
+        return [TinyKvFullCache()]
+
+    def __call__(self, inputs, offsets, cache, logits_to_keep=1):
+        offset = offsets[0] if isinstance(offsets, list) else int(offsets)
+        self.calls.append((offset, inputs.shape[1]))
+        length = inputs.shape[1]
+        key = mx.zeros((1, 1, length, 1), dtype=mx.float32)
+        cache[0].update_and_fetch(key, key)
+        logits = mx.zeros((1, 1, 4), dtype=mx.float32)
+        return logits.at[..., 1].add(1)
 
 
-def assert_causal_mask_faster_than_all_zero_mask(
-    stream: mx.Stream,
-    batch: int,
-    h_q: int,
-    h: int,
-    l: int,
-    s: int,
-    e: int,
-    scale: float = 0.9,
-    precision=mx.float32,
-):
-    q_shape = (batch, h_q, l, e)
-    kv_shape = (batch, h, s, e)
-    mask_shape = (batch, h_q, l, s)
+def test_chunked_prefill_bounds_work_and_advances_cache():
+    model = FakeModel()
+    request = Request(model, FakeTokenizer(), "1234567", prefill_max_step=3)
 
-    with mx.stream(stream):
-        query = mx.random.uniform(shape=q_shape, dtype=precision)
-        key = mx.random.uniform(shape=kv_shape, dtype=precision)
-        value = mx.random.uniform(shape=kv_shape, dtype=precision)
-        zero_mask = mx.zeros(shape=mask_shape, dtype=mx.float32)
+    request.try_prefill()
+    assert request.offset == 3
+    assert request.kv_cache[0].offset == 3
+    assert not request.is_prefill_done
 
-        for _ in range(3):
-            mx.eval(flash_attention(query, key, value, scale=scale, mask="causal"))
-            mx.eval(flash_attention(query, key, value, scale=scale, mask=zero_mask))
+    request.try_prefill()
+    assert request.offset == 6
+    assert request.kv_cache[0].offset == 6
+    assert not request.is_prefill_done
 
-    causal_samples = []
-    zero_mask_samples = []
-    for round_idx in range(6):
-        if round_idx % 2 == 0:
-            causal_samples.append(
-                time_flash_attention(
-                    stream, query, key, value, scale=scale, mask="causal"
-                )
-            )
-            zero_mask_samples.append(
-                time_flash_attention(
-                    stream, query, key, value, scale=scale, mask=zero_mask
-                )
-            )
-        else:
-            zero_mask_samples.append(
-                time_flash_attention(
-                    stream, query, key, value, scale=scale, mask=zero_mask
-                )
-            )
-            causal_samples.append(
-                time_flash_attention(
-                    stream, query, key, value, scale=scale, mask="causal"
-                )
-            )
+    request.try_prefill()
+    assert request.offset == 7
+    assert request.kv_cache[0].offset == 7
+    assert request.is_prefill_done
+    assert request.next_token == 1
+    assert model.calls == [(0, 3), (3, 3), (6, 1)]
 
-    causal_time = median(causal_samples)
-    zero_mask_time = median(zero_mask_samples)
-    assert causal_time < zero_mask_time, (
-        "Expected causal mask to be faster than an all-zero mask, got "
-        f"causal={causal_time:.6f}s and zero_mask={zero_mask_time:.6f}s."
-    )
-
-
-@pytest.mark.parametrize("mask_mode", ["no_mask", "mask", "causal"])
-def test_task_2_flash_attention_cpu_small(mask_mode: str):
-    attention_helper(mx.cpu, 6, 3, 2, 5, 3, 1, mask_mode)
-
-
-@pytest.mark.parametrize("mask_mode", ["no_mask", "mask"])
-def test_task_2_flash_attention_cpu(mask_mode: str):
-    attention_helper(mx.cpu, 18, 6, 7, 5, 3, 10, mask_mode)
-
-
-@pytest.mark.parametrize("mask_mode", ["no_mask", "mask", "causal"])
-def test_task_2_flash_attention_cpu_large(mask_mode: str):
-    attention_helper(mx.cpu, 28, 4, 16, 128, 16, 3, mask_mode)
-
-
-def test_task_2_flash_attention_cpu_causal_mask_faster_than_all_zero_mask():
-    assert_causal_mask_faster_than_all_zero_mask(
-        stream=mx.cpu,
-        batch=1,
-        h_q=8,
-        h=8,
-        l=128,
-        s=128,
-        e=128,
-    )
-
-
-@pytest.mark.parametrize("mask_mode", ["no_mask", "mask"])
-def test_task_3_flash_attention_gpu_extra_small(mask_mode: str):
-    attention_helper(mx.gpu, 1, 1, 5, 7, 4, 1, mask_mode)
-
-
-@pytest.mark.parametrize("mask_mode", ["no_mask", "mask", "causal"])
-def test_task_3_flash_attention_gpu_small(mask_mode: str):
-    attention_helper(mx.gpu, 6, 3, 2, 5, 3, 1, mask_mode)
-
-
-@pytest.mark.parametrize("mask_mode", ["no_mask", "mask"])
-def test_task_3_flash_attention_gpu(mask_mode: str):
-    attention_helper(mx.gpu, 18, 6, 7, 5, 3, 10, mask_mode)
-
-
-@pytest.mark.parametrize("mask_mode", ["no_mask", "mask", "causal"])
-def test_task_3_flash_attention_gpu_large(mask_mode: str):
-    attention_helper(mx.gpu, 28, 4, 16, 128, 16, 3, mask_mode, precision=mx.bfloat16)
-
-
-@pytest.mark.parametrize("mask_mode", ["no_mask", "mask", "causal"])
-def test_task_3_flash_attention_gpu_d128_partial_tiles(mask_mode: str):
-    attention_helper(mx.gpu, 8, 2, 35, 128, 47, 1, mask_mode, precision=mx.bfloat16)
-
-
-def test_task_3_flash_attention_gpu_causal_mask_faster_than_all_zero_mask():
-    assert_causal_mask_faster_than_all_zero_mask(
-        stream=mx.gpu,
-        batch=2,
-        h_q=8,
-        h=8,
-        l=512,
-        s=512,
-        e=128,
-        precision=mx.bfloat16,
-    )
+    with pytest.raises(ValueError, match="after done"):
+        request.try_prefill()
